Modernize the code: Remove iVars that are referenced by properties. Rename properties that begin with "is" to not begin with "is", but have the getter still begin with "is". This is in keeping with modern Cocoa practices. Add a synthesizer for the cache property in MainView.

+++ Classes/TouchInfoStore.m
@@ -34,8 +34,8 @@
 	return self;
 }
 
-- (int) count {
-	return (int) currentTouchInfos.count;
+- (NSInteger) count {
+	return currentTouchInfos.count;
 }
 
 - (void) storeTouches:(NSSet *)touches {
